chatbot/requirements.txt
 
flask==2.0.1
flask-login==0.5.0
flask-mysqldb==0.2.0
google-generativeai==0.3.0
python-dotenv==0.19.0
Werkzeug==2.0.1